# server/docs/swagger/swagger.yaml
basePath: /
definitions:
  config.Bitaxe:
    properties:
      enabled:
        type: boolean
      fallbackPort:
        type: integer
      fallbackUrl:
        type: string
      fallbackUser:
        type: string
      hostname:
        type: string
      ip:
        type: string
      mac:
        description: |-
          Mac is the device's MAC address (colons/hyphens optional, normalized
          by StorageKey) -- the storage-directory key, stable across IP/
          location changes unlike Ip. Populated either by hand or by the network
          discovery flow (see internal/discovery), which reads it straight off
          the device's own /api/system/info response.
        type: string
      model:
        type: string
      poolSchedule:
        items:
          $ref: '#/definitions/config.CronSchedule'
        type: array
      port:
        type: integer
      url:
        type: string
      user:
        type: string
    type: object
  config.CronSchedule:
    properties:
      cron:
        type: string
      target:
        $ref: '#/definitions/config.PoolTarget'
    type: object
  config.PoolTarget:
    enum:
    - primary
    - fallback
    type: string
    x-enum-varnames:
    - Primary
    - Fallback
  handler.AlertEntry:
    properties:
      alerts:
        items:
          $ref: '#/definitions/model.Alert'
        type: array
      hostname:
        type: string
      minerIp:
        type: string
      minerMac:
        type: string
      timestamp:
        type: string
    type: object
  handler.AlertEpisode:
    properties:
      firstSeen:
        type: string
      hostname:
        type: string
      lastSeen:
        type: string
      message:
        type: string
      minerIp:
        type: string
      minerMac:
        type: string
      occurrences:
        type: integer
      peakValue:
        type: number
      threshold:
        type: number
      type:
        type: string
    type: object
  handler.AlertHistoryResponse:
    properties:
      episodes:
        items:
          $ref: '#/definitions/handler.AlertEpisode'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  handler.StatsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.MinerInfo'
        type: array
      total:
        type: integer
    type: object
  handler.bitaxesResponse:
    properties:
      bitaxes:
        items:
          $ref: '#/definitions/config.Bitaxe'
        type: array
      lastUpdated:
        type: string
    type: object
  model.Alert:
    properties:
      message:
        type: string
      threshold:
        description: the threshold in effect at calculation time, omitted otherwise
        type: number
      type:
        description: '"tempHigh", "fanHigh", "offline", "macMismatch", "firmwareUpdate"'
        type: string
      value:
        description: the reading that triggered it (temp/fan), omitted otherwise
        type: number
    type: object
...
  model.MinerInfo:
    properties:
      alerts:
        description: |-
          Alerts is whatever the feeder computed and persisted at the poll that
          produced this snapshot (temp/fan out of range, offline, mac mismatch,
          firmware update) -- the current state as of Timestamp, not a live value.
        items:
          $ref: '#/definitions/model.Alert'
        type: array
      alive:
        description: Health check
        type: boolean
      aliveCheckedAt:
        type: string
      bestDiff:
        description: Highest difficulty value ever solved by the miner since start‑up
        type: integer
      blockFound:
        type: integer
      deviceModel:
        description: Computed device model (e.g. "NerdQAxe++", "Bitaxe 602")
        type: string
      electricityRatePerKwh:
        description: Electricity cost
        type: number
      energyJPerTh:
        description: Energy consumption per terahash (Joules per TH)
        type: number
      error:
        description: |-
          Error surfaces a serious inconsistency the UI should call out
          distinctly from a normal offline/unreachable state -- currently just
          a mismatch between the configured mac: and what the device itself
          reports (wrong device at this IP, or a config typo).
        type: string
      fallbackStratumDashboardURL:
        description: Web dashboard URL for the fallback pool
        type: string
      fallbackStratumURL:
        type: string
      fallbackStratumUser:
        type: string
      fanspeed:
        description: Fan speed (linked to the Temp field)
        type: number
      hashRateTHs:
        description: Additional fields requested by the API
        type: number
      hostname:
        description: Human-readable name from config
        type: string
      ip:
        description: Miner IP address (derived from the folder name)
        type: string
      isUsingFallbackStratum:
        type: integer
      latestVersion:
        description: Latest firmware version available on GitHub
        type: string
      macAddr:
        description: Mac address
        type: string
      networkDifficulty:
        description: Current network difficulty reported by the pool (unitless)
        type: integer
      power:
        description: Watts (Joules per second)
        type: number
      releaseURL:
        description: GitHub releases page for the latest firmware version
        type: string
      responseTime:
        description: Ping/latency to the pool (ms).
        type: number
      sharesAccepted:
        type: integer
      sharesRejected:
        type: integer
      stratumDashboardURL:
        description: Web dashboard URL for the primary pool
        type: string
      stratumURL:
        description: Pool urls
        type: string
      stratumUser:
        description: Username (typically miner ID) for the primary pool
        type: string
      temp:
        description: Primary temperature sensor reading (°C)
        type: number
      timestamp:
        type: string
      totalSharesAccepted:
        type: integer
      totalSharesRejected:
        type: integer
      totalUptimeSeconds:
        description: |-
          Total* are persistent, reboot-surviving counters (see
          internal/storage.Totals) -- unlike SharesAccepted/SharesRejected and
          UptimeSeconds above, which reset whenever the device itself reboots,
          these keep accumulating for as long as the miner has been tracked.
          Zero until the feeder has written at least one totals.json for this
          miner, or until the backfill tool has run.
        type: integer
      updateAvailable:
        description: True when LatestVersion != Version
        type: boolean
      uptimeSeconds:
        description: Miner uptime in seconds
        type: integer
      version:
        description: Firmware version (e.g. "v2.12.2")
        type: string
      voltage:
        description: Volts (millivolts)
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Local, read/write REST API for AxeOs-compatible miners (Bitaxe, NerdAxe) on your LAN.
    No authentication — internal LAN use only.

    Endpoints tagged "dashboard-api" run on this binary (default :8080).
    Endpoints tagged "remote-dashboard-api" run on the separate remote-dashboard-api binary (default :8081),
    a read-only API serving miners pushed to hashboard.live — see the project README.
  title: axeos-dashboard API
  version: "1.0"
paths:
...
  /api/config/discover:
    get:
      description: Probes a single IP (?ip=) or sweeps a CIDR range (?cidr=, max /24,
        defaults to the server's own local /24 subnet when both are omitted). ip takes
        priority if both are set.
      parameters:
      - description: Probe exactly this one IP instead of scanning a range (e.g. a
          device on a different VLAN than the server). Takes priority over cidr.
        in: query
        name: ip
        type: string
      - description: 'Network range to sweep, e.g. ''192.168.1.0/24'' (max /24 = 254
          hosts). Omitted: auto-detects and scans the server''s own local /24 subnet.
          Ignored if ip is set.'
        in: query
        name: cidr
        type: string
      - description: Per-host probe timeout, e.g. '500ms' or '2s' (default 1s, capped
          at 3s). Raise it and retry if a scan/probe found nothing on a slow network.
        in: query
        name: timeout
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.bitaxesResponse'
        "400":
          description: invalid/too-large cidr, or invalid/too-large timeout
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: no AxeOS device found at the given ip
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Discover AxeOS devices on the network
      tags:
      - dashboard-api
  /api/config/miners:
    get:
      description: Returns every miner entry in the managed miners config, including
        disabled ones. Read-only -- see POST/DELETE for writing.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.bitaxesResponse'
      summary: List configured miners
      tags:
      - dashboard-api
    post:
      consumes:
      - application/json
      description: Upserts one or more miner entries into the managed miners.yml,
        matched by MAC. New entries are always saved enabled.
      parameters:
      - description: Miner entries to add/update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.bitaxesResponse'
      produces:
      - application/json
      responses:
        "200":
          description: the full, updated list of configured miners
          schema:
            $ref: '#/definitions/handler.bitaxesResponse'
        "400":
          description: invalid request body or entry
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: server not started with -miners -- nowhere safe to write
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: failed to write the miners file
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Add or update configured miners
      tags:
      - dashboard-api
...
  /api/miners/{hostnameOrIp}/restart:
    post:
      description: Restarts the device via its AxeOs restart endpoint.
      parameters:
      - description: Miner IP or configured hostname
        in: path
        name: hostnameOrIp
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: miner not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Restart a miner
      tags:
      - dashboard-api
  /api/miners/{hostnameOrIp}/stats:
    get:
      description: Returns today's JSONL entries (one per poll cycle) for a single
        miner, used by the "Today's History" chart.
      parameters:
      - description: Miner IP or configured hostname
        in: path
        name: hostnameOrIp
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatsResponse'
        "404":
          description: miner not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get today's stats for one miner
      tags:
      - dashboard-api
  /api/miners/alerts:
    get:
      description: Returns each miner's N most recent alert-bearing entries (today's
        history only).
      parameters:
      - description: max entries to return per miner (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AlertEntry'
            type: array
      summary: List the most recent miner alerts
      tags:
      - dashboard-api
  /api/miners/alerts/history:
    get:
      description: Returns every alert episode (consecutive same-type occurrences
        collapsed into one) on the given day across all miners, most recently active
        first, optionally filtered by ip/type.
      parameters:
      - description: the single day to scan (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: page number, 1-based (default 1)
        in: query
        name: page
        type: integer
      - description: episodes per page (default 50, max 200)
        in: query
        name: pageSize
        type: integer
      - description: filter to this miner IP only
        in: query
        name: ip
        type: string
      - description: filter to this alert type only
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AlertHistoryResponse'
      summary: List a single day's recorded miner alerts, grouped into episodes and
        paginated
      tags:
      - dashboard-api
  /api/miners/pool/fallback/enable:
    put:
      description: Switches miner(s) to the primary or fallback pool and restarts
        them to apply it.
      parameters:
      - description: Miner IP or hostname — omit to apply to all configured miners
        in: query
        name: miner
        type: string
      responses:
        "204":
          description: No Content
      summary: Switch stratum pool
      tags:
      - dashboard-api
  /api/miners/pool/primary/enable:
    put:
      description: Switches miner(s) to the primary or fallback pool and restarts
        them to apply it.
      parameters:
      - description: Miner IP or hostname — omit to apply to all configured miners
        in: query
        name: miner
        type: string
      responses:
        "204":
          description: No Content
      summary: Switch stratum pool
      tags:
      - dashboard-api
  /api/miners/set/wifi:
    put:
      description: Pushes the configured WiFi SSID/password to miner(s) and restarts
        them. Disabled unless wifi.on is set in config.
      parameters:
      - description: Miner IP or hostname — omit to apply to all configured miners
        in: query
        name: miner
        type: string
      responses:
        "204":
          description: No Content
        "405":
          description: WiFi updates disabled by config
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update WiFi credentials
      tags:
      - dashboard-api
swagger: "2.0"